Log send include (keybase) system log
- Move platform specific log stuff into install
- Don't show warning on log file not found, if the web don't have the log, then it wasn't there
