--- go/client/cmd_log_send_nix.go
@@ -1,10 +1,0 @@
-// Copyright 2015 Keybase, Inc. All rights reserved. Use of
-// this source code is governed by the included BSD license.
-
-// +build !windows
-
-package client
-
-func GetInstallLogPath() (string, error) {
-	return "", nil
-}
