+++ go/install/install_osx.go
@@ -1016,3 +1016,13 @@ func RunApp(context Context, log Log) error {
 	}
 	return nil
 }
+
+// InstallLogPath doesn't exist on darwin as an independent log file (see desktop app log)
+func InstallLogPath() (string, error) {
+	return "", nil
+}
+
+// SystemLogPath is where privileged keybase processes log to on darwin
+func SystemLogPath() string {
+	return "/Library/Logs/keybase.system.log"
+}
